+movedOrdersInBottomBar&AdjustedSetupAuthModalButtons
