+++ app/(tabs)/_layout.tsx
@@ -14,7 +14,6 @@ const styles = createStyles(StyleSheet);
 // Set to true to enable the tab, false to disable (but keep visible)
 const TAB_ENABLED = {
   WALLET: false, // Change to true to enable Wallet tab
-  PHONE: false, // Change to true to enable Phone tab
 };
 
 // Disabled tab styling
@@ -91,8 +90,28 @@ export default function TabLayout() {
         }}
       />
       <Tabs.Screen
+        name={ROUTE_NAMES.ORDERS}
+        options={{
+          title: "Orders",
+          headerShown: true,
+          header: () => (
+            <SafeAreaView edges={["top"]}>
+              <Header title="Orders" style={{ justifyContent: "center" }} />
+            </SafeAreaView>
+          ),
+          tabBarIcon: ({ color, focused }) => (
+            <TabBarIcon
+              name={focused ? "receipt" : "receipt-outline"}
+              color={color}
+              style={styles.tabBarIcon}
+            />
+          ),
+        }}
+      />
+      <Tabs.Screen
         name={ROUTE_NAMES.PHONE}
         options={{
+          href: null, // Hide from tab bar — moved into Settings as disabled "Contact"
           title: "Contacts",
           headerShown: true,
           header: () => (
@@ -100,24 +119,6 @@ export default function TabLayout() {
               <Header title="Contacts" style={{ justifyContent: "center" }} />
             </SafeAreaView>
           ),
-          tabBarIcon: ({ color, focused }) => (
-            <TabBarIcon
-              name={focused ? "call" : "call-outline"}
-              color={TAB_ENABLED.PHONE ? color : Theme.colors.inactive}
-              style={[
-                styles.tabBarIcon,
-                !TAB_ENABLED.PHONE && { opacity: DISABLED_TAB_OPACITY },
-              ]}
-            />
-          ),
-        }}
-        // NOTE: Remove when tab is enabled
-        listeners={{
-          tabPress: (e) => {
-            if (!TAB_ENABLED.PHONE) {
-              e.preventDefault();
-            }
-          },
         }}
       />
       <Tabs.Screen
