+++ app/(tabs)/settings.tsx
@@ -1,11 +1,10 @@
-import React, { useState, useCallback, useEffect, useMemo } from "react";
+import React, { useState, useCallback, useMemo } from "react";
 import {
   StyleSheet,
   FlatList,
   TouchableOpacity,
   View,
   ScrollView,
-  ActivityIndicator,
   // Switch,
   Text,
   Switch,
@@ -21,20 +20,6 @@ import { Ionicons } from "@expo/vector-icons";
 import { useKokio } from "@/hooks/useKokio";
 import { useAuthRelay } from "@/hooks/useAuthRelayer";
 import { SafeAreaView } from "react-native-safe-area-context";
-import { useRouter } from "expo-router";
-import * as Clipboard from "expo-clipboard";
-import type { StoredPurchasedESIM } from "@/providers/kokioProvider";
-import { getAllEsims, getEsim } from "@/utils/bff/esim";
-import type { ESimDocument } from "@/utils/bff/esim";
-import { getOrderStatus } from "@/utils/bff/order";
-import type { OrderStatusResponse } from "@/utils/bff/order";
-import { labelForStatus, colorForStatus } from "@/utils/orderStatus";
-import ESIMItem from "@/components/ESIMItem";
-
-type EnrichedOrder = StoredPurchasedESIM & {
-  liveEsim?: ESimDocument;
-  liveStatus?: OrderStatusResponse;
-};
 
 const createStyles = () => StyleSheet.create({
   container: {
@@ -68,31 +53,6 @@ const createStyles = () => StyleSheet.create({
   },
   iconRight: {
     marginLeft: 16,
-  },
-  orderCardWrapper: {
-    marginBottom: 4,
-  },
-  orderMeta: {
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    alignItems: 'center',
-    gap: 6,
-    paddingHorizontal: 4,
-    paddingBottom: 8,
-  },
-  orderStatusBadge: {
-    borderRadius: 8,
-    paddingVertical: 3,
-    paddingHorizontal: 8,
-  },
-  orderStatusText: {
-    fontSize: 11,
-    fontWeight: '600',
-    textTransform: 'uppercase',
-  },
-  orderMetaText: {
-    fontSize: 12,
-    color: Theme.colors.muted,
   },
   aboutContainer: {
     borderRadius: 25,
@@ -163,7 +123,7 @@ const createStyles = () => StyleSheet.create({
 });
 
 const MENU_ITEM_ENABLED = {
-  PROFILE: true,
+  CONTACT: false, // moved from the bottom Phone tab — enable once contacts feature is ready
   PRIVACY: false, // enable once privacy policy is ready
 };
 
@@ -216,205 +176,6 @@ const MenuItem = ({
       />
     </View>
   </TouchableOpacity>
-  );
-};
-
-const CopyRow = ({ label, value }: { label: string; value: string }) => {
-  const [copied, setCopied] = useState(false);
-  const handleCopy = useCallback(async () => {
-    await Clipboard.setStringAsync(value);
-    setCopied(true);
-    setTimeout(() => setCopied(false), 1500);
-  }, [value]);
-  return (
-    <TouchableOpacity onPress={handleCopy} style={{ flexDirection: 'row', alignItems: 'center', marginTop: 6 }}>
-      <Text style={{ color: Theme.colors.mutedForeground, fontSize: 11, flex: 1 }} numberOfLines={1}>
-        {label}: <Text style={{ color: Theme.colors.foreground }}>{value}</Text>
-      </Text>
-      <Ionicons name={copied ? 'checkmark' : 'copy-outline'} size={14} color={copied ? Theme.colors.success : Theme.colors.mutedForeground} style={{ marginLeft: 6 }} />
-    </TouchableOpacity>
-  );
-};
-
-const OrderCard = ({ order, onInstall }: { order: EnrichedOrder; onInstall?: (lpa: string) => void }) => {
-  const { isDark } = useTheme();
-  const styles = useMemo(createStyles, [isDark]);
-  const [expanded, setExpanded] = useState(false);
-  const { eSimItem, transactionData, liveEsim, liveStatus } = order;
-  const statusColor = colorForStatus(transactionData.orderStatus);
-
-  const lpa = liveEsim?.smdpAddress && liveEsim?.matchingId
-    ? `LPA:1$${liveEsim.smdpAddress}$${liveEsim.matchingId}`
-    : transactionData.installationDetails?.qrcode ?? null;
-
-  const invoiceUrl = liveStatus?.stripeInvoiceUrl ?? null;
-  const flagged = liveStatus?.flaggedForManualReview ?? false;
-  const supportRef = transactionData.correlationId ?? transactionData.orderId;
-
-  return (
-    <TouchableOpacity onPress={() => setExpanded(e => !e)} activeOpacity={0.85} style={styles.orderCardWrapper}>
-      <ESIMItem item={eSimItem} showBuyButton={false} />
-      <View style={styles.orderMeta}>
-        <View style={{ flexDirection: 'row', alignItems: 'center', gap: 8 }}>
-          {transactionData.orderStatus ? (
-            <View style={[styles.orderStatusBadge, { backgroundColor: statusColor + '22' }]}>
-              <Text style={[styles.orderStatusText, { color: statusColor }]}>
-                {labelForStatus(transactionData.orderStatus)}
-              </Text>
-            </View>
-          ) : null}
-          {flagged ? (
-            <View style={[styles.orderStatusBadge, { backgroundColor: Theme.colors.goldenYellow + '22' }]}>
-              <Text style={[styles.orderStatusText, { color: Theme.colors.goldenYellow }]}>Under Review</Text>
-            </View>
-          ) : null}
-          <Ionicons name={expanded ? 'chevron-up' : 'chevron-down'} size={14} color={Theme.colors.mutedForeground} style={{ marginLeft: 'auto' }} />
-        </View>
-
-        {expanded ? (
-          <View style={{ marginTop: 10, gap: 2 }}>
-            {transactionData.iccid ? <CopyRow label="ICCID" value={transactionData.iccid} /> : null}
-            {supportRef ? <CopyRow label="Ref" value={supportRef} /> : null}
-            {transactionData.paymentMethod ? (
-              <Text style={{ color: Theme.colors.mutedForeground, fontSize: 11, marginTop: 6 }}>
-                Payment: {transactionData.paymentMethod}
-              </Text>
-            ) : null}
-            {invoiceUrl ? (
-              <TouchableOpacity onPress={() => Linking.openURL(invoiceUrl)} style={{ marginTop: 8 }}>
-                <Text style={{ color: Theme.colors.highlight, fontSize: 12 }}>View invoice →</Text>
-              </TouchableOpacity>
-            ) : null}
-            {lpa ? (
-              <>
-                <CopyRow label="LPA" value={lpa} />
-                {onInstall ? (
-                  <TouchableOpacity onPress={() => onInstall(lpa)} style={{ marginTop: 8 }}>
-                    <Text style={{ color: Theme.colors.highlight, fontSize: 12 }}>Install eSIM →</Text>
-                  </TouchableOpacity>
-                ) : null}
-              </>
-            ) : null}
-            {liveEsim?.planHistory?.length ? (
-              <View style={{ marginTop: 10 }}>
-                <Text style={{ color: Theme.colors.mutedForeground, fontSize: 11, marginBottom: 4 }}>Plan history</Text>
-                {liveEsim.planHistory.map((entry, i) => (
-                  <Text key={i} style={{ color: Theme.colors.foreground, fontSize: 11, marginBottom: 2 }}>
-                    {entry.planId}  ·  {entry.validity}d  ·  {new Date(entry.purchaseDate).toLocaleDateString()}
-                  </Text>
-                ))}
-              </View>
-            ) : null}
-          </View>
-        ) : null}
-      </View>
-    </TouchableOpacity>
-  );
-};
-
-const OrdersContent = ({ orders, onClose, onInstall }: { orders: StoredPurchasedESIM[]; onClose: () => void; onInstall: (lpa: string) => void }) => {
-  const { isDark } = useTheme();
-  const styles = useMemo(createStyles, [isDark]);
-  const [enrichedOrders, setEnrichedOrders] = useState<EnrichedOrder[]>(orders);
-  const [isFetching, setIsFetching] = useState(true);
-
-  useEffect(() => {
-    let cancelled = false;
-    const fetchLive = async () => {
-      setIsFetching(true);
-      try {
-        const liveEsims = await getAllEsims();
-        const liveMap: Record<string, ESimDocument> = {};
-        liveEsims.forEach(e => { liveMap[e.esimId] = e; });
-
-        const merged = await Promise.all(orders.map(async (order): Promise<EnrichedOrder> => {
-          const { transactionData } = order;
-
-          // For orders with a known esimId, attach live ESimDocument
-          if (transactionData.esimId && liveMap[transactionData.esimId]) {
-            const live = liveMap[transactionData.esimId];
-            const liveStatus = transactionData.correlationId
-              ? await getOrderStatus(transactionData.correlationId).catch(() => undefined)
-              : undefined;
-            return {
-              ...order,
-              liveEsim: live,
-              liveStatus,
-              transactionData: {
-                ...transactionData,
-                iccid: live.iccid ?? transactionData.iccid,
-                orderStatus: live.activationStatus ?? transactionData.orderStatus,
-                planId: live.planId ?? transactionData.planId,
-                installationDetails: live.installationDetails ?? transactionData.installationDetails,
-              },
-            };
-          }
-
-          // For pending/processing orders, re-fetch status via correlationId
-          if (transactionData.correlationId && !transactionData.installationDetails?.qrcode) {
-            try {
-              const latest = await getOrderStatus(transactionData.correlationId);
-              const liveEsim = latest.esimId
-                ? liveMap[latest.esimId] ?? await getEsim(latest.esimId).catch(() => undefined)
-                : undefined;
-              return {
-                ...order,
-                liveEsim,
-                liveStatus: latest,
-                transactionData: {
-                  ...transactionData,
-                  orderId: latest.orderId || transactionData.orderId,
-                  iccid: latest.iccid ?? transactionData.iccid,
-                  orderStatus: latest.orderStatus ?? transactionData.orderStatus,
-                  esimId: latest.esimId ?? transactionData.esimId,
-                  installationDetails: liveEsim?.installationDetails ?? latest.installationDetails ?? transactionData.installationDetails,
-                },
-              };
-            } catch {
-              return order;
-            }
-          }
-
-          return order;
-        }));
-
-        if (!cancelled) setEnrichedOrders(merged);
-      } catch {
-        // live fetch failed — show stored data as-is
-      } finally {
-        if (!cancelled) setIsFetching(false);
-      }
-    };
-    fetchLive();
-    return () => { cancelled = true; };
-  }, []);
-
-  return (
-    <View style={styles.aboutContainer}>
-      <View style={styles.aboutHeader}>
-        <ThemedText style={styles.aboutTitle}>Orders</ThemedText>
-        <TouchableOpacity onPress={onClose} style={styles.closeButton}>
-          <Ionicons name="close-outline" size={28} color="white" />
-        </TouchableOpacity>
-      </View>
-      {isFetching ? (
-        <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-          <ActivityIndicator size="large" color={Theme.colors.secondary} />
-        </View>
-      ) : enrichedOrders.length === 0 ? (
-        <ThemedText style={[styles.aboutText, { textAlign: 'center', marginTop: 32 }]}>
-          No orders yet.
-        </ThemedText>
-      ) : (
-        <FlatList
-          data={enrichedOrders}
-          keyExtractor={(_, i) => i.toString()}
-          renderItem={({ item }) => <OrderCard order={item} onInstall={onInstall} />}
-          showsVerticalScrollIndicator={false}
-          contentContainerStyle={{ paddingBottom: 8 }}
-        />
-      )}
-    </View>
   );
 };
 
@@ -515,12 +276,10 @@ const ContactContent = ({ onClose }: { onClose: () => void }) => {
 
 export default function MenuScreen() {
   const { logout } = useAuthRelay();
-  const { kokio, clearKokioUser } = useKokio();
-  const router = useRouter();
+  const { clearKokioUser } = useKokio();
   const { isDark, toggleTheme } = useTheme();
 
   const [showAbout, setShowAbout] = useState(false);
-  const [showOrders, setShowOrders] = useState(false);
   const [showContact, setShowContact] = useState(false);
   const bg = useThemeColor({}, "background");
   const styles = useMemo(createStyles, [isDark]);
@@ -528,11 +287,10 @@ export default function MenuScreen() {
   const menuItems = [
     {
       id: "1",
-      title: "Orders",
-      iconLeft: "receipt-outline",
+      title: "Contact",
+      iconLeft: "call-outline",
       iconRight: "chevron-forward-outline",
-      disabled: !MENU_ITEM_ENABLED.PROFILE,
-      action: () => setShowOrders(true),
+      disabled: !MENU_ITEM_ENABLED.CONTACT,
     },
     {
       id: "3",
@@ -577,16 +335,7 @@ export default function MenuScreen() {
   return (
     <SafeAreaView style={{ flex: 1, backgroundColor: bg }}>
       <ThemedView style={styles.container}>
-        {showOrders ? (
-          <OrdersContent
-            orders={kokio.purchasedESIMs}
-            onClose={() => setShowOrders(false)}
-            onInstall={(lpa) => {
-              setShowOrders(false);
-              router.push({ pathname: '/(tabs)/installation', params: { qrcode: lpa } });
-            }}
-          />
-        ) : showAbout ? (
+        {showAbout ? (
           <AboutContent onClose={() => setShowAbout(false)} />
         ) : showContact ? (
           <ContactContent onClose={() => setShowContact(false)} />
