+++ constants/route.constants.ts
@@ -4,6 +4,7 @@ export const ROUTE_NAMES = {
   CHECKOUT: "checkout/[id]",
   WALLET: "(wallet)",
   PHONE: "phone",
+  ORDERS: "orders",
   SETTINGS: "settings",
   BY_COUNTRY: "country/[id]",
   BY_REGION: "region/[id]",
@@ -32,7 +33,7 @@ export const TAB_BAR_ENABLED_ROUTES: string[] = [
   `${ROUTE_NAMES.SHOP}/${ROUTE_NAMES.HOME}/Regions`,
   `${ROUTE_NAMES.SHOP}/${ROUTE_NAMES.HOME}/Global`,
   ROUTE_NAMES.WALLET,
-  ROUTE_NAMES.PHONE,
+  ROUTE_NAMES.ORDERS,
   ROUTE_NAMES.SETTINGS,
   `${ROUTE_NAMES.WALLET}/${ROUTE_NAMES.TOKENS}`,
   `${ROUTE_NAMES.WALLET}/${ROUTE_NAMES.TRANSACTIONS}`,
